Add test for renderer option
