+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
         name: isort (python)
 
 -   repo: https://github.com/psf/black
-    rev: 20.8b1
+    rev: 22.3.0
     hooks:
       - id: black
         language_version: python3
