+++ test/util.py
@@ -47,3 +47,25 @@ def first(iterator: Iterator[T]) -> Optional[T]:
         return next(iterator)
     except StopIteration:
         return None
+
+
+BUSY_WAIT_SCRIPT = """
+import time, sys
+
+def do_nothing():
+    pass
+
+def busy_wait(duration):
+    end_time = time.time() + duration
+
+    while time.time() < end_time:
+        do_nothing()
+
+def main():
+    print('sys.argv: ', sys.argv)
+    busy_wait(0.25)
+
+
+if __name__ == '__main__':
+    main()
+"""
